Factorize PollViewModel to SharedViewModel
